requirements.txt
openai>=1.0.0
langchain>=0.3.0
faiss-cpu
python-dotenv